chore: address issue #617
