+++ app/services/refraction_static_artifacts/__init__.py
@@ -5,6 +5,11 @@ from __future__ import annotations
 from app.services.refraction_static_artifacts import contract
 from app.services.refraction_static_artifacts import _legacy
 from app.services.refraction_static_artifacts._legacy import *  # noqa: F403
+from app.services.refraction_static_artifacts.line_profile import (
+    build_refraction_line_profile_qc_arrays as build_refraction_line_profile_qc_arrays,
+    build_refraction_line_profile_qc_payload as build_refraction_line_profile_qc_payload,
+    write_refraction_line_profile_qc_artifacts as write_refraction_line_profile_qc_artifacts,
+)
 
 for _name in contract.__all__:
     globals()[_name] = getattr(contract, _name)
